Moved some other types and snackbar to its own folder
