+++ src/components/shared-theme/AppTheme.tsx
@@ -1,21 +1,12 @@
 import React, { Fragment, useMemo } from "react"
 import { ThemeProvider, createTheme } from "@mui/material/styles"
-import type { ThemeOptions } from "@mui/material/styles"
 import { inputsCustomizations } from "./customizations/inputs"
 import { dataDisplayCustomizations } from "./customizations/dataDisplay"
 import { feedbackCustomizations } from "./customizations/feedback"
 import { navigationCustomizations } from "./customizations/navigation"
 import { surfacesCustomizations } from "./customizations/surfaces"
 import { colorSchemes, typography, shadows, shape } from "./themePrimitives"
-
-interface AppThemeProps {
-  children: React.ReactNode;
-  /**
-   * This is for the docs site. You can ignore it or remove it.
-   */
-  disableCustomTheme?: boolean;
-  themeComponents?: ThemeOptions["components"];
-}
+import { AppThemeProps } from "../../types/common"
 
 export default function AppTheme(props: AppThemeProps) {
   const { children, disableCustomTheme, themeComponents } = props
