+++ src/types/snackbar/types.ts
@@ -4,3 +4,9 @@ export type SnackbarProps = {
   severity: "error" | "warning" | "info" | "success";
   onClose: () => void;
 }
+
+export type SnackbarState = {
+  open: boolean;
+  message: string;
+  severity: "error" | "warning" | "info" | "success";
+}
